Add class FEImmersedSurfaceValues
That has the same shape function values/derivatives as FEValues,
but has JxW-values corresponding to integration over a surface
immersed in the cell. This is achieved by passing an
ImmersedSurfaceQuadrature instead of a Quadrature to the constructor.

Add a virtual function fill_fe_immersed_surface_values on Mapping to
be able to compute the correct JxW-values and surface normals. Make it
throw ExcNotImplemented by default, but override it with an
implementation on MappingCartesian.