CMake: fix "Release" and "Debug" only configure for CMake 3.30
Beginning with version 3.30 CMake makes it an error to specify a
nonexistent target in a call to target_link_libraries() even if that
target is never used internally and only part of the exported interface.
This is a problem as we want to add configuration dependent targets with
a generator expression that are later exported:

  $<$<CONFIG:Release>:dealii::dealii_release> (...)

Note that we cannot simply use $<$<CONFIG:Release>:dealii_release> as
CMake then does not complete "dealii_release" to "dealii::dealii_release"
upon target export...

As a workaround we create alias targets with the "dealii::" namespace,
which makes the call to target_link_libraries() succeed.